mythshutdown: rename commandlineparser.* to mythshutdown_…
